授权问题

+++ hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/service/hjh/impl/HjhTenderServiceImpl.java
@@ -1212,7 +1212,7 @@ public class HjhTenderServiceImpl extends BaseTradeServiceImpl implements HjhTen
 
         String roleIsOpen = systemConfig.getRoleIsopen();
         if(StringUtils.isNotBlank(roleIsOpen) && roleIsOpen.equals("true")){
-            if (userInfo.getRoleId() != 1) {
+            if (userInfo.getRoleId().intValue() != 1) {
                 throw new CheckException(MsgEnum.ERR_AMT_TENDER_ONLY_LENDERS);
             }
         }
